1-concurrent_coroutines.py

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -8,6 +8,7 @@
     The list of the delays should be in ascending order without
     using sort() because of concurrency.
 """
+
 import asyncio
 from typing import List
 
@@ -15,7 +16,7 @@ from typing import List
 wait_random = __import__("0-basic_async_syntax").wait_random
 
 
-async def wait_n(n: int, max_delay: float) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """wait_n function"""
     delay = await asyncio.gather(
         *tuple(map(lambda _: wait_random(max_delay), range(n)))
